.github/workflows: update NodeJS to 20
With all dependency upgrade I've done, the support for NodeJS 12 is now
broken. So it's time to move on to a newer, still supported version.

NodeJS 20 is choosen because a.) while it's not an LTS yet, it will
become an LTS soon, and b.) it matches the version in netdev/devenv
Docker image.